requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# LangGraph / LangChain
langgraph==0.0.35
langchain==0.3.28
langchain-community==0.3.31
langchain-openai==0.1.0
langchain-google-genai==2.0.0
langchain-core==0.1.33

# Data models
pydantic==2.6.4
pydantic-settings==2.2.1

# LLM / AI
groq==0.5.0
tiktoken==0.6.0

# HTTP / Scraping
aiohttp==3.9.3
playwright==1.42.0

# Cache utilities
cachetools==5.3.3

# Storage
redis==5.0.3
supabase==2.4.2  # Optional: only needed if SUPABASE_URL/KEY are configured

# Resilience
tenacity==8.2.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
